+++ NEWS.md
@@ -64,6 +64,11 @@ Note: We're only listing outstanding class updates.
       ineffect inside its body, without affecting the original Proc.
       [[Feature #22097]]
 
+* Range
+
+    * `Range#clamp` is added. It returns a new `Range` instance whose
+      begin and end values are clamped to the given bounds. [[Feature #22175]]
+
 * Regexp
 
     * All instances of `Regexp` are now frozen, not just literals.
@@ -245,6 +250,7 @@ A lot of work has gone into making Ractors more stable, performant, and usable. 
 [Feature #21981]: https://bugs.ruby-lang.org/issues/21981
 [Feature #22137]: https://bugs.ruby-lang.org/issues/22137
 [Feature #22139]: https://bugs.ruby-lang.org/issues/22139
+[Feature #22175]: https://bugs.ruby-lang.org/issues/22175
 [Feature #22185]: https://bugs.ruby-lang.org/issues/22185
 [PR #17201]: https://github.com/ruby/ruby/pull/17201
 [RubyGems-v4.0.4]: https://github.com/rubygems/rubygems/releases/tag/v4.0.4